// src/models/knowledge-base-data-files.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger Data Provider
 * Документация API Провайдера Данных
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface KnowledgeBaseDataFiles
 */
export interface KnowledgeBaseDataFiles {

    /**
     * File name
     *
     * @type {string}
     * @memberof KnowledgeBaseDataFiles
     */
    name: string;

    /**
     * Url for file downloading
     *
     * @type {string}
     * @memberof KnowledgeBaseDataFiles
     */
    url: string;
}